Move dekaf utils to the connector utils file
