+++ src/components/materialization/useGenerateCatalog.ts
@@ -43,7 +43,7 @@ import {
 import { FormStatus } from 'stores/FormState/types';
 import { useSourceCaptureStore } from 'stores/SourceCapture/Store';
 import { DekafConfig } from 'types';
-import { isDekafConnector } from 'utils/misc-utils';
+import { isDekafConnector } from 'utils/connector-utils';
 import { encryptEndpointConfig } from 'utils/sops-utils';
 import { generateTaskSpec } from 'utils/workflow-utils';
 import { ConnectorConfig } from '../../../deps/flow/flow';
